refactor(frontend): extract MapCanvas type declarations → MapCanvas.types.ts
U1 of the MapCanvas consolidation epic (#884). The three top-level
type-only declarations leave MapCanvas.tsx for a dedicated module:

  - MapCanvasProps (the prop contract; carries the App.tsx scope/camera/
    artboard wiring doc-comments #736/#740/#760/#761/#782)
  - SelectedObsState (the #718 popover-projection state)
  - ResolvedAdaptiveData (the Concern-B leafCache payload, epic #539)

Why: these erase at build (zero runtime), so co-locating them with the
2,800-line imperative core only adds noise to the file the rest of the
epic shrinks. Pulling them out is the lowest-risk Wave-0 cut and gives
later units a stable home for shared prop/state types.

Behavior-preserving: MapCanvas.tsx back-imports all three and re-exports
MapCanvasProps, so the rendered output and the module's public surface
are unchanged. The orphaned type imports the move left behind
(FamilySilhouette, AdaptiveTile, ResolvedGrid) are pruned from
MapCanvas.tsx and now resolve in MapCanvas.types.ts instead.

knip stays clean: the new file's exports are all consumed by the
back-import, and the re-exported MapCanvasProps is covered by
ignoreExportsUsedInFile (its in-file destructuring consumer).

Type-only move — gated by tsc/build (no new runtime test). Local gate:
frontend build + 1178 tests + lint + knip all green.

Closes #885. Part of #884.
